refactor: getCategories only in server component
