+++ components/layout/category-nav/category-chip.tsx
@@ -1,20 +1,19 @@
 "use client";
 
 import Link from "next/link";
+import { usePathname } from "next/navigation";
 import { CATEGORY_ALL, CATEGORY_NAME_LABELS } from "@/constants/category";
 import Chip from "./chip";
 
 type Props = {
   name: string;
   count: number;
-  currentCategoryName: string;
 };
 
-export default function CategoryChip({
-  name,
-  count,
-  currentCategoryName,
-}: Props) {
+export default function CategoryChip({ name, count }: Props) {
+  const pathname = usePathname();
+  const currentCategoryName = getCategoryNameFromPath(pathname);
+
   const urlPath = getPath(name);
   const label = getLabel(name, count);
   const isSelected = name === currentCategoryName;
@@ -24,6 +23,13 @@ export default function CategoryChip({
       <Chip label={label} isSelected={isSelected} />
     </Link>
   );
+}
+
+function getCategoryNameFromPath(pathname: string) {
+  const regex = /pages|\/|category|\d+/g;
+  const category = pathname.replace(regex, "");
+
+  return category === "" ? CATEGORY_ALL : category;
 }
 
 function getPath(name: string) {
