+++ components/layout/category-nav/index.tsx
@@ -1,13 +1,8 @@
-"use client";
-
-import { usePathname } from "next/navigation";
 import { CATEGORY_ALL } from "@/constants/category";
 import { getAllCategoriesWithCount } from "@/lib/posts";
 import CategoryChip from "@/components/layout/category-nav/category-chip";
 
 export default function CategoryNav() {
-  const pathname = usePathname();
-  const currentCategoryName = getCategoryNameFromPath(pathname);
   const categories = getAllCategoriesWithCount();
 
   return (
@@ -15,18 +10,10 @@ export default function CategoryNav() {
       {categories?.map((category) => (
         <CategoryChip
           key={category.name}
-          currentCategoryName={currentCategoryName}
           name={category.name}
           count={category.count}
         />
       ))}
     </div>
   );
-}
-
-function getCategoryNameFromPath(pathname: string) {
-  const regex = /pages|\/|category|\d+/g;
-  const category = pathname.replace(regex, "");
-
-  return category === "" ? CATEGORY_ALL : category;
 }
